+++ tests/custom-discriminator/tests/custom-discriminator.ts
@@ -48,7 +48,7 @@ describe("custom-discriminator", () => {
         "confirmed"
       );
       const myAccount = await program.account.myAccount.fetch(
-        pubkeys.myAccount
+        pubkeys.myAccount as anchor.web3.PublicKey
       );
       assert.strictEqual(myAccount.field, field);
     });