+++ tests/safety-checks/test.sh
@@ -6,7 +6,7 @@ echo "Building programs"
 # Build the UncheckedAccount variant.
 #
 pushd programs/unchecked-account/
-output=$(anchor build 2>&1 > /dev/null)
+output=$(anchor build --ignore-keys 2>&1 > /dev/null)
 if ! [[ $output =~ "Struct field \"unchecked\" is unsafe" ]]; then
    echo "Error: expected /// CHECK error"
    exit 1
@@ -17,7 +17,7 @@ popd
 # Build the AccountInfo variant.
 #
 pushd programs/account-info/
-output=$(anchor build 2>&1 > /dev/null)
+output=$(anchor build --ignore-keys 2>&1 > /dev/null)
 if ! [[ $output =~ "Struct field \"unchecked\" is unsafe" ]]; then
    echo "Error: expected /// CHECK error"
    exit 1
@@ -28,7 +28,7 @@ popd
 # Build the control variant.
 #
 pushd programs/ignore-non-accounts/
-if ! anchor build ; then
+if ! anchor build --ignore-keys ; then
    echo "Error: anchor build failed when it shouldn't have"
    exit 1
 fi